initial commit for #2480 build python driver without protoc
This replaces the use of the binary protoc with a Python script to translate the .proto file into a useable form. In order to make the translator easy to write and compatible with the JavaScript and Ruby this adds another hierarchy layer to most of the calls, so those are changed as well.

Additionally this changed the build process for all three drivers to use this new tool.